feat: extract SVG quality checks into diagram-testkit package
Extract test_svg_quality.py into an installable Python package at
packages/diagram-testkit/. No logic changes — identical behavior
verified by running all 8 fixture SVGs through the checks.

Package structure:
- geometry.py: BBox, bbox_from_path_d, text_bbox, path_to_linestring
- parser.py: ParsedSVG dataclass, parse_svg()
- checks.py: 6 check functions + run_all_checks()
- cli.py: diagram-testkit lint file1.svg file2.svg

Key changes:
- check_data_layer_position generalized to check_cluster_alignment
  with src/dst cluster name parameters
- run_all_checks() runs only universal checks (not cluster alignment)
- Original test_svg_quality.py is now a thin wrapper importing from
  diagram_testkit

17/17 tests pass. All 8 fixture SVGs still fail at least one check.
