+++ packages/diagram-testkit/pyproject.toml
@@ -1,0 +1,23 @@
+[project]
+name = "diagram-testkit"
+version = "0.1.0"
+description = "SVG quality linter for generated diagrams"
+requires-python = ">=3.10"
+license = "MIT"
+dependencies = [
+    "svgpathtools",
+    "shapely",
+]
+
+[project.optional-dependencies]
+test = ["pytest"]
+
+[project.scripts]
+diagram-testkit = "diagram_testkit.cli:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/diagram_testkit"]
