+++ packages/diagram-testkit/src/diagram_testkit/__main__.py
@@ -1,0 +1,5 @@
+"""Allow running as `python -m diagram_testkit`."""
+
+from .cli import main
+
+main()
