+++ packages/diagram-testkit/tests/test_cli.py
@@ -1,0 +1,34 @@
+"""CLI smoke tests."""
+
+import subprocess
+import sys
+from pathlib import Path
+
+FIXTURES_DIR = Path(__file__).parent / "fixtures"
+
+
+def _run_cli(*args: str) -> subprocess.CompletedProcess:
+    return subprocess.run(
+        [sys.executable, "-m", "diagram_testkit", *args],
+        capture_output=True,
+        text=True,
+    )
+
+
+def test_cli_lint_bad_fixture_exits_1():
+    bad_svgs = list(FIXTURES_DIR.glob("architecture-arrow-*.svg"))
+    assert bad_svgs, "No arrow-crossing fixture found"
+    result = _run_cli("lint", str(bad_svgs[0]))
+    assert result.returncode == 1
+    assert "FAIL" in result.stdout
+
+
+def test_cli_lint_nonexistent_file_exits_1():
+    result = _run_cli("lint", "/nonexistent/file.svg")
+    assert result.returncode == 1
+    assert "File not found" in result.stdout
+
+
+def test_cli_no_command_exits_1():
+    result = _run_cli()
+    assert result.returncode == 1
